+++ MicrosoftTeams.php
@@ -184,6 +184,8 @@ class MicrosoftTeams extends \Piwik\Plugin
             throw new \Exception(Piwik::translate('MicrosoftTeams_IncomingWebhookInvalidErrorMessage'));
         }
 
+        $this->assertWebhookDestinationAllowed($parameters[self::MS_TEAMS_INCOMING_WEBHOOK_URL_PARAMETER]);
+
         $parameters[self::MS_TEAMS_INCOMING_WEBHOOK_URL_PARAMETER] = htmlspecialchars_decode($parameters[self::MS_TEAMS_INCOMING_WEBHOOK_URL_PARAMETER]);
     }
 
@@ -419,6 +421,8 @@ class MicrosoftTeams extends \Piwik\Plugin
                 throw new \Exception(Piwik::translate('MicrosoftTeams_IncomingWebhookInvalidErrorMessage'));
             }
 
+            $this->assertWebhookDestinationAllowed($parameters[self::MS_TEAMS_INCOMING_WEBHOOK_URL_PARAMETER]);
+
             $parameters[self::MS_TEAMS_INCOMING_WEBHOOK_URL_PARAMETER] = htmlspecialchars_decode($parameters[self::MS_TEAMS_INCOMING_WEBHOOK_URL_PARAMETER]);
         }
     }
@@ -432,6 +436,33 @@ class MicrosoftTeams extends \Piwik\Plugin
         $host = trim($host, '[]');
 
         return filter_var($host, FILTER_VALIDATE_IP) !== false || ctype_digit($host);
+    }
+
+    private function assertWebhookDestinationAllowed(string $webhookUrl): void
+    {
+        $host = parse_url($webhookUrl, PHP_URL_HOST);
+        if (empty($host)) {
+            return;
+        }
+
+        $host = trim($host, '[]');
+
+        if ($this->isLocalMatomoHost($host)) {
+            throw new \Exception(Piwik::translate('MicrosoftTeams_IncomingWebhookInvalidErrorMessage'));
+        }
+    }
+
+    private function isLocalMatomoHost(string $host): bool
+    {
+        $host = strtolower($host);
+
+        if (in_array($host, ['localhost', 'localhost.localdomain', 'ip6-localhost'], true)) {
+            return true;
+        }
+
+        $matomoHost = parse_url(SettingsPiwik::getPiwikUrl(), PHP_URL_HOST);
+
+        return !empty($matomoHost) && strcasecmp(trim($matomoHost, '[]'), $host) === 0;
     }
 
     /**